requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.2.0

# Local LLM + image generation
torch>=2.4.0
transformers>=4.45.0
accelerate>=0.34.0
bitsandbytes>=0.43.0
diffusers>=0.30.0
safetensors>=0.4.0
huggingface_hub>=0.24.0

# Media
moviepy>=1.0.3,<2
Pillow>=10.4.0
numpy>=2.0.0

# Audio IO
soundfile>=0.12.0
pyttsx3>=2.90
PyQt6>=6.7.0
